fix: resolve Next.js build error by moving ProjectCard to a shared component
